TDD with Withdrawals:
We will use Test Driven Design (TDD) for the withdraw function.  This will allow money to be removed from an account and it the opposite of a deposit.

You must:

 Run all the unit tests and get them all green. Two AccountsUnitTests will fail and a number of AccountServicesUnit Tests will fail.  Thesae are all associated with Withdrawal.
Fix the Withdrawal() method in Account Services so it remove money from the account.
This might require changes to the account class.
Some accounts get reward points when you deposit money into them, but you should not get any reward points when you withdraw money and add a negative amount.
Get the Withdrawal button working on the user interface now that you have the account services working.