Refactor or Bug Fix
I chose the following option:
- Refactor the code so that withdrawals use their own methods in Accounts and don't just add a negative amount.

Once you have all the unit tests green (passing), you must select some part of the code to refactor or select a bug to fix.

Select ONE thing to fix.  Here are some possibilities, though you can also find your own:

Refactor the code so that withdrawals use their own methods in Accounts and don't just add a negative amount.
Update the GUI so that when you select an account, the account balance is updated.
Add some error checking for one of the fields in the GUI
Allow the user to select the account type when creating a new account and don't always create Silver accounts.